services/backend/requirements.txt
alembic==1.1.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
# torch==1.2.0
# torchvision==0.4.0
urllib3==1.25.3
Werkzeug==0.15.6
psycopg2-binary==2.8.3